bot/requirements.txt
python-telegram-bot
requests
redis
python-dotenv
pyyaml